shift_planner/requirements.txt
streamlit==1.40.0
streamlit-authenticator==0.4.1
pandas>=2.2.0
numpy>=2.0.0
plotly>=5.24.0
Pillow>=10.4.0
openpyxl>=3.1.5
PyYAML>=6.0.2
pulp>=2.7.0           # optional; MILP optimizer will auto-fallback if missing
python-dotenv>=1.0.1
python-dateutil>=2.9.0
openai>=1.54.0       # AI-powered scheduling optimization
dspy-ai>=2.4.0       # Structured LLM programming with signatures
pydantic>=2.0.0      # Data validation and type safety
